move run_instance to eval.py
